JobWebsite/requirements.txt
aniso8601>=4.0.1
asn1crypto>=0.24.0
certifi>=2018.11.29
cffi>=1.11.5
chardet>=3.0.4
Click>=7.0
coverage>=4.5.2
cryptography>=2.4.2
Django>=2.1.4
django-bower>=5.2.0
django-crispy-forms>=1.7.2
django-nose>=1.4.6
django-seed>=0.1.9
djangorestframework>=3.9.0
djangorestframework-jwt>=1.11.0
Faker>=1.0.1
flake8>=3.6.0
idna>=2.8
itsdangerous>=1.1.0
Jinja2>=2.10
MarkupSafe>=1.1.0
mccabe>=0.6.1
nose>=1.3.7
ntlm-auth>=1.2.0
passlib>=1.7.1
Pillow>=5.3.0
psycopg2-binary>=2.7.6.1
pycodestyle>=2.4.0
pycparser>=2.19
pyflakes>=2.0.0
python-dateutil>=2.7.5
pytz>=2018.7
six>=1.12.0
text-unidecode>=1.2
urllib3>=1.24.1
Werkzeug>=0.14.1
django-tables2>=2.0.5
django-import-export>= 1.2.0
